Added the ability to hide item nametags if the stack size is one.
Added the ability to hide the stack size if the stack size is one.
Items will now show a colorless version of their custom display name (If available) in their nametag.

The name format for items has been default changed to "&f{TYPE} [&6{AMT}x]". The square brackets are used to remove and show the amount when needed.