Do entity collection in Bukkit's main thread for StackingTask
Paper-Spigot got more strict in 1.19.2 about calling
stuff async.

I do not know what the whole StackingTask does and
don't feel like cleaning it up completely is possible right now.

So I only moved the getEntities-calls to the main
thread and have the async task wait for it.
Not great but it fixed the bug.


SD-9374
SD-9377
SD-9392
SD-9401